Upgrade c-ares library to 1.34.4 (#38559)
Upgrades the c-ares DNS library to 1.34.4. This is to gain access to
`ares_reinit()` feature.

This is to work towards solving issue #34785.

@agrawroh Already attempted to solve this, but I think ran into local
testing issues which were unrelated to the version of c-ares.

I'm going to resurrect @agrawroh's PR (#37011) with this version of
c-ares and attempt to get it across the finish line.

---------

Signed-off-by: Clif Houck <me@clifhouck.com>
Signed-off-by: Ryan Northey <ryan@synca.io>