wasm: fix possible crash when resetting context after vm panic (#38865)
If the wasm plugin send a http call out and panic before the response.
Then when the response is arrived, the clean up in the
`addAfterVmCallAction` will bypassed. Then, a dirty pending request
pointer will be kept in the Context object.

When the Context is destroyed because configuration update or vm reload,
the the dangling pointer will crash envoy.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: added.
Platform Specific Features: n/a.

---------

Signed-off-by: wangbaiping(wbpcode) <wbphub@gmail.com>