fix: credential injector init manager (#38798)
fix:
https://github.com/envoyproxy/gateway/pull/5496#discussion_r1998304248

We should use the initManager in the DualInfo because the Credential
Injector can be used for both HCM filter and upstream filter. Using the
initManger from the ServerFactoryContext for HCM filter causes the
secret to be added to the server initManager when it's already in the
initialized state.

Change log should not be required as this fixes a bug introduced in [a
RP](https://github.com/envoyproxy/envoy/pull/38398) that just merged
after v1.33.0 .

@yanavlasov

---------

Signed-off-by: Huabing (Robin) Zhao <zhaohuabing@gmail.com>