oauth2: clear extraneous cookies and don't set refresh cookie when refresh flow not enabled (#38941)
This addresses two problems primarily concerned with the refresh token
flow. First, when refresh tokens are disabled by configuration, if the
Authorization Server (AS) returns a refresh token, Envoy will dutifully
set it as a response cookie and include it in the HMAC. This is not an
obvious bug, but it certainly unexpected, especially when coupled with
the second issue.

The second issue is described in #38919. tl;dr: If there is a
`RefreshToken` cookie, but the AS doesn't send back a refresh token on
authorization, the HMAC will continually fail and the only way to fix
this is to delete the cookie.

This PR makes it so that:
1. If refresh tokens are disabled by configuration, don't set the cookie
2. If any particular token is set in a cookie, but not returned by the
AS, actively clear the cookie.
1. The exception here is that if we went through the refresh flow, but
the AS didn't return a new refresh token, just use the existing refresh
token cookie. It's unfortunately common for some Authorization Servers
to not rotate refresh tokens upon access token refresh
2. One issue here is that in a scenario like #33943 where the IdToken
isn't send during the token refresh, this PR changes it so that the
IdToken cookie will be cleared, whereas before it would be kept.
Technically this fixes #33943, but it also clears the IdToken, which
might be surprising to people (though i suspect that it will surprise
nobody currently since it is broken right now!)

Commit Message: oauth2: clear extraneous cookies and don't set refresh
cookie when refresh flow not enabled
Additional Description:
Risk Level: low
Testing: unit testing and manual testing ensuring that the repro from
the issue no longer repros. note that i am not able to manually test the
case where a refresh token is not returned during token refresh because
i don't have an Authorization Server with this configuration
Docs Changes:
Release Notes: Fixes case where an existing RefreshToken cookie will
break oauth token flow if the authorization server does not return a
refresh token.
Platform Specific Features:
Fixes: #38919 and #33943

---------

Signed-off-by: Ben Plotnick <bplotnick@netflix.com>