route match: add filter state (#38842)
adding a route matcher on filter state. This will allow to match on
filter state in the router. Filter state allows for example to match on
IP CIDR which is not the case for dynamic metadata and it contains
information that are not available in the metadata.

This was already done for the matcher tree in the route configuration:
https://github.com/envoyproxy/envoy/issues/32807 and
https://github.com/envoyproxy/envoy/pull/33175

Commit Message: route match: add filter state
Additional Description:
Risk Level: Low
Testing: Test
Docs Changes: No
Release Notes:
Platform Specific Features:
[Optional Runtime guard:]
[Optional Fixes #Issue]
[Optional Fixes commit #PR or SHA]
[Optional Deprecated:]
[Optional [API
Considerations](https://github.com/envoyproxy/envoy/blob/main/api/review_checklist.md):]

Signed-off-by: Florent Lecoultre <florent.lecoultre@datadoghq.com>