Switch back to static linking for coverage tests
This is a workaround for the LLVM issue that affects coverage (see
https://github.com/llvm/llvm-project/issues/32849).

TL;DR: basically dynamic linking + definitions of functions in the
headers result in mismatching checksums for coverage counters which in
turn results in incorrect coverage reports.

The effects of this issue are not very deterministic (e.g., when the
issue happens is affected by way to many factors to reliably predict or
avoid the issue), that's why we haven't noticed it before, but during
clang-18 migration the issue caused some coverage failures.

Switching to static linking should workardound the issue completely.
Before migration to clang-18 we couldn't do it, because coverage tests
just didn't build because we hit some relocation overflows. However
since version 17 LLVM lld changed the order of sections in the produced
binaries, which made relocations overflows way less likely in the
code model used by default (which I think is "small").

NOTE: we could aslo try and switch to large code model, but not only it
will result in slower code, I also don't think it's as well tested and
supported as small code model, so we probably should not.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>