dfp: Don't filter IP address families from the DNS cache
In Mobile, whenever there is a network change, there's a check to see if
there is IPv6 connectivity on the current network. If there is no IPv6
connectivity and the `dns_cache_set_ip_version_to_remove` runtime guard
is true, the IPv6 addresses would get filtered out from the DNS lookup
response.

This is problematic because if there is a network change and the device
now has IPv6 connectivity, the DNS cache will have filtered out the IPv6
addresses already and won't be able to use them to connect to the peer.

This change fixes the issue by storing all resolved IP addresses in the
DNS cache, and only filtering out the IP address families that don't
have connectivity in the DFP cluster. This has the further benefit of
avoiding having to re-resolve DNS host entries whenever there is a
network change.

Signed-off-by: Ali Beyad <abeyad@google.com>