ext_proc: fix coverage reports for ext_proc tests (#39041)
Commit Message:

In #38898 @botengyao added an exception for the ext_proc tests for
coverage. That's because clang source based coverage reported missing
coverage for the comments (the desirable behavior here is that comments
are not counted in the coverage reports).

Digging through it, what I realized is that for only some tests report
counters for the comments. And all the tests that report thos bogus
counters don't actually exercise the affected code.

I still don't fully understand what triggers the issue here, but that
observation suggests a workaround that this PR implements. I basically
remove dependency on the grpc_client_impl.h from the tests that don't
need it.

I started a thread on LLVM discussions
(https://discourse.llvm.org/t/llvm-source-based-coverage-sometimes-generates-counter-for-comments/85730),
but given that I couldn't come up with a small enough reproducer, I
don't think that we can resolve it quickly.

That being said, removing unnecessary includes and moving functions from
headers to cc files is in general a good strategy as it reduces
depndencies between translation units and reduces the amount of work
compiler has to do, so I think we can still push this workaround
upstream, even though I don't have a root cause for the problem.

Additional Description:

It was reported in #37911, but other than that I don't think it's
relevant.

Risk Level: n/a
Testing: regular CI tests, but with coverage targets changed to remove
the exception for the ext_proc.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>