Downgrade connection-ID and worker-index mismatch from assert to error. (#39016)
Commit Message: Downgrade connection-ID and worker-index mismatch from
assert to error when kernel_worker_routing_ is true.
Additional Description: kernel_worker_routing is not enabled atomically
on the server side. If there is a QUIC client talking to server during
the kernel_worker_routing setup, the cid and worker-index mismatch
assertion can trigger, which is not desirable. If packet routing is not
working as intended, the new error will show up in logs.
Risk Level: LOW

---------

Signed-off-by: Haoyue Wang <haoyuewang@google.com>