fips: update boringssl to use the main branch (#39017)
Change-Id: I23c1fcf05b4713d848a2f46852525d6c2506986a

Commit Message: Update the BoringSSL FIPS build to use the same commit
as a regular non-FIPS build (9/13/2024) per the revised FedRAMP policy.

Additional Description: BoringSSL considers the main branch to be the
"update stream" in the policy. The new recommendation is to use the
latest stable versions of the compilers, so we also bump Clang version
to be the same as the regular build (from 14 to 18). Please note that
this change **does not** include setting the compliance policy, as that
would be a breaking change.

Risk Level: low (as we match the thoroughly tested non-FIPS build)
Testing: regression tests passed
Issue: https://github.com/envoyproxy/envoy/issues/38353
Docs Changes: none
Release Notes: yes (it is quite plausible a default might have changed
in the crypto settings due to the large version bump from 2022 to 2024,
so this is a notable change).

---------

Signed-off-by: Kuat Yessenov <kuat@google.com>