network: add close through network filter_manager support to check filter's close status (#38778)
To solve the data loss issue during the close sequence especially for
the middle external processing network filter in the filter chain.
#38779


![image](https://github.com/user-attachments/assets/c1422fee-d749-4ce6-90dc-9cb6b963e4f4)


This PR will let the close manage by filter manager for some cases, and
also introduce an enum `StopIterationAndDontClose` for network filter,
which a `continueClosing()` is expected if the filter manger is closing
the connection. It will check all the networks' filter status before
closing.

### Alternative Design
To provide more flexibility, we can split the disable close and
`StopIteration` functionalities, and introduce an disableClose(bool) to
change a filter's `pending_close_` status.

### Key changes in **ConnectionImpl**

For the read path directly from socket:

- The `RemoteClose` will close the socket when the socket is closed when
half close is not enabled or both ways are half closed if half close is
enabled immediately. With this PR right now the `RemoteClose` will be
managed by the filter manager. Since a filter could return
`StopIterationAndDontClose`, the `closeThroughFilerManager()` will pause
the close until the filter calls `continueClosing`.

For the write path directly from socket:
- If it is in the `FlushWrite` or `FlushWriteAndDelay` status that is
triggered by `close()`, the socket will close immediately since the
close() is managed by the filter manager.
- For half close cases, if both ways are half closed, the close will
also be managed by the filter manager, since the read filter is possible
in the close pending status.

For the `close()` that is initialized by Envoy's connection:

- We will only let the close manage by the filter manger if it is
`FlushWrite` and `FlushWriteAndDelay`
- `Abort`, `NoFlush`, and `AbortReset` will still close the socket
immediately.
- We still want to have the ability to flush the write buffer after the
wait, and that is the major reason that we don't directly manage the
`closeSocket(LocalClose)`.

For remote early close:

- If half-close is enabled, this is never activated.
- If half-close is disabled, there are two scenarios when the close
event is monitored:
1. During the closeInternal(_), which comes from filter manager already.
2. When an early close is detected while the connection is
read-disabled. It is okay to bypass the filter manager's status since
there will be data loss even in normal cases.

### Key changes in Filter_manager
- Introduce a `StopIterationAndDontClose` for filter status and
`conntinueClosing()`
- Introduce the `state_` in filter manager to check the pending close
status
- It supports multiple filter use cases by introducing a counter.

**Note**:

- this will not impact normal traffic without the
`StopIterationAndDontClose`.
- If the connection is pending close, only
`injectWriteDataToFilterChain` and `injectReadDataToFilterChain` is
called.
- The close is tracked by counting if all network filters have finished
the job.


Commit Message:
Additional Description:
Risk Level: Low for normal case, and High if feature is used.
Testing:
Docs Changes:
Release Notes:
Platform Specific Features:

---------

Signed-off-by: Boteng Yao <boteng@google.com>