Disable exceptions thrown by the json parser (#39029)
The current implementation uses the `accept`
function to check the validity of the json string before parsing it.
This PR removes the call to `accept`, disables the exceptions thrown by
the `parse` function and checks if `parse` retuns a discarded json
object for the validity of the json string.
Risk Level: n/a
Testing: No extra testing needed, the current tests are sufficient to
test this change.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: numanelahi <numanelahi@google.com>