Disable exceptions in spdlog (#39045)
This should have been done originally, since this is a risk on the
dataplane if input to spdlog causes an exception. It is better to
produce a truncated log output than crash.

Risk Level: Low, we do not rely on exceptions in spdlog
Testing: unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Yan Avlasov <yavlasov@google.com>