Drop FIPS build of IPP
The reason we had FIPS and non-FIPS build of IPP is because we used to
use different versions of BoringSSL for FIPS and non-FIPS builds - the
FIPS build being older.

IPP needed a patch to make it work with an older version of BoringSSL,
so that's why we had a separate IPP build for FIPS.

Since we now moved BoringSSL FIPS to use the update stream we were able
to update the version of BoringSSL we use in FIPS build and so we don't
need a patch for IPP anymore.

This PR removes the separate FIPS build of IPP and fixes contrib build
in FIPS mode.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>