Fix local FIPS builds of BoringSSL in Envoy
There is a bad interaction between Bazel and Go file embedding.
Basically, Go file embedding does not allow embedding non-regular
files (like symlinks), but when building things locally with Bazel
it instead of copying all the required files into the directory
where the action is executed creates symbolic links instead, which
breaks Go.

We worked around known cases of this issue in the BoringSSL FIPS
build script by copying files that we know are embedded and therefore
will break Bazel build. However, after updating BoringSSL version it
seems like we now have some new cases of this issues and this PR tries
to address that.

NOTE: There was a proper workaround implemented for this on Go side that
would not require us to play whackamole (
https://go-review.googlesource.com/c/go/+/647315), but it was only
recently submitted and haven't made its way into Go release yet
(probably the next Go release, 1.25 will have this included and we would
be able to tell Go to follow symbolic links).

NOTE: There is also a way to tell Bazel to create hard links instead of
symbolic links, but that comes with some caveats that I don't fully
understand and therefore will not suggest.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>