dynamic_modules: prevent reentrance on local reply (#39206)
Commit Message: dynamic_modules: prevent reentrance on local reply
Additional Description:

LocalReply of the stream callbacks can cause a reentrance. For example,
it causes encodeHeaders to be called inline on top of the call to
LocalReply. That causes a few problems e.g. multiple mutable references
in Rust. In practice, modules won't need these callbacks to be called
again since at the callsite of LocalReply, they already know all the
details. This adds a boolean flag to prevent that and added a test that
would fail otherwise. Note that this behavior matches the one of Wasm
extension.

Risk Level: low
Testing: done
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features:

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>