fips: detect validation and defaults at runtime (#39192)
Change-Id: I3d6bb4a1e21bf992bfeaef64d53640101a67e04b

Commit Message: Detect FIPS at runtime and select the code path in the
defaults and the validation based on that. This addresses a
contradictory build where a FIPS boringssl is linked but the
corresponding macro BORINGSSL_FIPS is not set. In such cases, FIPS
validations are not applied where they should have been, so this PR
fixes this.
Risk Level: medium, may change defaults on custom builds
Testing: done
Docs Changes: none
Release Notes: none (official builds are OK)

---------

Signed-off-by: Kuat Yessenov <kuat@google.com>