Drop fuzz_coverage target from CI
I'm trying to switch back to static linking for coverage tests in Envoy
CI to work around a bug in Clang/LLVM (see
https://github.com/llvm/llvm-project/issues/32849).

However the change that switches to static linking fails `fuzz_coverage`
with what seems like some kind of resource constraint (e.g., linker gets
killed, I suspect, as a result of OOM, since static linking requires
more resources).

@yanavalsov suggested that fuzz_coverage target might not be the best
way to measure fuzzing coverage (see
https://github.com/envoyproxy/envoy/pull/39030#issuecomment-2807912729)
and given that I think we should at least consider disabling it until we
can make it better. That's what this PR does.

NOTE: @phlax also mention that he did some work to migrate coverage to
EngFlow builds in the past and there might be PR that we could finish
that will migrate fuzz_coverage to EngFlow. There is a chance that it
will address the issue - I'm mentioning it here to give a complete
overview of the possible alternatives here.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>