hcm: Ensure operations are not called on deleted stream decoders
Previously, it was possible for the `ActiveStream` in the
`HttpConnectionManager` to get deleted while still trying to process
packets from the codec.

This change uses the `ActiveStreamHandle` with weak pointer semantics to
ensure that even with incoming data packets, methods are not called on a
deleted `ActiveStream`, which represents the HCM's `RequestDecoder`.

Signed-off-by: Ali Beyad <abeyad@google.com>