Fix environment-dependent UnexpectedFormatCpuAllocatedLine test
LinuxContainerCpuStatsReader that this test exercises takes two file
paths (CPU allocation path and CPU times path) in constructor and
both of those construcotr arguments have default values that point to
some cgroup files that may or may not exist in the system where the
test runs.

When you don't explicitly provide both paths in the constructor in test
LinuxContainerCpuStatsReader will fallback to the default value and
will try to read and parse the actual cgroup CPU times file
/sys/fs/cgroup/cpu/cpuacct.usage and if the file does not exist it will
fail.

That's exactly the problem with UnexpectedFormatCpuAllocatedLine test.
When it creates LinuxContainerCpuStatsReader it explicitly provides a
path to a test CPU allocation file, but does not provide a path to a
test CPU times file (which is understandble give that the test does not
care about the content of the CPU times file).

However the implementation of LinuxContainerCpuStatsReader::getCpuTimes
is such that if it fails to read any of those files the functio returns
a failure.

This makes this test somewhat non-deterministic in the sense that the
lines of code that this test exercises depend on the environment and
whether /sys/fs/cgroup/cpu/cpuacct.usage exists.

This issue affected Envoy coverage tests when we tried to migrate them
from Google RBE backend to EngFlow. The same test worked fine and
produced 100% coverage for LinuxContainerCpuStatsReader on Google RBE
backends, but when we migrated to EngFlow the coverage changed and
triggered a coverage test failure.

This PR refactors the test a bit to make sure that both CPU allocation
CPU times files are always available and valid, unless the test
explicitly overrides it.

I would also argue that this change makes the tests a bit easier to read
as it removes the irrelvant boilerplate to a fixture constructor.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>