Fix environment-dependent UnexpectedFormatCpuAllocatedLine test (#39304)
Commit Message:
Tests were somewhat non-deterministic in the sense that the
lines of code that this test exercises depend on the environment and
whether /sys/fs/cgroup/cpu/cpuacct.usage exists.

This issue affected Envoy coverage tests when we tried to migrate them
from Google RBE backend to EngFlow. The same test worked fine and
produced 100% coverage for LinuxContainerCpuStatsReader on Google RBE
backends, but when we migrated to EngFlow the coverage changed and
triggered a coverage test failure.

This PR refactors the test a bit to make sure that both CPU allocation
CPU times files are always available and valid, unless the test
explicitly overrides it.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>