grid: Propagate transport failure reason of both connection pools (#39253)
Previously, if the TCP pool failed before the HTTP/3 pool failed, the
`ConnectivityGrid` would propagate the transport failure reason for both
the TCP and QUIC pools. However, if the QUIC pool failed first, the
`ConnectivityGrid` would not propagate the transport failure reason for
the QUIC pool (only for the TCP pool).

This change ensures that the `ConnectivityGrid` propagates the transport
failure reason for *both* TCP and QUIC pools if both pools fail in the
`ConnectivityGrid`, regardless of the order in which the pools fail.

---------

Signed-off-by: Ali Beyad <abeyad@google.com>