Link test binaries statically when run coverage
Static linking helps to work around the issue with LLVM/Clang
source-based coverage (see
https://github.com/llvm/llvm-project/issues/32849).

This PR switches to static linking for just test coverage and does not
do the same for fuzz coverage. That's because Envoy CI is hitting some
resource constraints when building fuzz targets with coverage
instrumentation.

We will fix that by striping the fuzz targets of some unncessary
dependencies and switching to EngFlow backend for coverage builds and
tests, but that will require addressing a couple of hard to understand
issues, so, for now, I'm just switching the coverage tests to static
linking and will follow up with the fuzz tests later once we addressed
all the blockers and switched to EngFlow backend for coverage.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>