Link test binaries statically when run coverage (#39354)
Static linking helps to work around the issue with LLVM/Clang
source-based coverage (see
https://github.com/llvm/llvm-project/issues/32849). On the flip side,
the coverage build and test run will take a bit longer (e.g., around
30m) with this change.

This PR switches to static linking for just test coverage and does not
do the same for fuzz coverage. That's because Envoy CI is hitting some
resource constraints when building fuzz targets with coverage
instrumentation.

We will fix that by striping the fuzz targets of some unncessary
dependencies and switching to EngFlow backend for coverage builds and
tests, but that will require addressing a couple of hard to understand
issues, so, for now, I'm just switching the coverage tests to static
linking and will follow up with the fuzz tests later once we addressed
all the blockers and switched to EngFlow backend for coverage.

Additional Description:

https://github.com/envoyproxy/envoy/pull/39030 provides some context for
EngFlow migration and https://github.com/envoyproxy/envoy/issues/39248
is a tracking bug.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>