conn pool: fix issue that can lead to insufficient connections (#39310)
If a connection starts draining while it has negative unused capacity
(which happens if a SETTINGS frame reduces allowed concurrency to below
the current number of requests), that connections unused capacity will
be included in total pool capacity even though it is unusable because it
is draining. This can result in not enough connections being established
for current pending requests.

This is most problematic for long-lived requests (such as streaming gRPC
requests or long-poll requests) because a connection could be in the
draining state for a long time.

Maybe fixes: #39238

Signed-off-by: Greg Greenway <ggreenway@apple.com>
(cherry picked from commit e44c4ebba793c340a81b18dd340f71db7e68f41d)