(header_mutation): support trimming trailers
As discussed in https://github.com/envoyproxy/envoy/issues/39382
with `ORCA` we have the grpc upstream sending load report in the
response trailers. This can be propagated by envoy to the client which
can lead to the client load balancing against envoy based on upstream
load reports. This is not ideal in scenarios where envoy acts a gateway
(e.g. cluster ingress in k8s).

This changes extends the header mutation filter to support mutating response trailers.

Notes on the implementation I wanted some feedback on:
- I didnt add support for trimming the request trailers. I think this can be easily
be added for parity but I cant think of a use case where this would be useful. So I
was a bit hesitant to add it proactively. Have to implement it based on feedback.
- I didnt extend the `header_mutation_integration_test` because they seemed to focus on
shared functionality between the header and trailer mutation. Happy to add it if you think
it would be useful.

Signed-off-by: sotiris <sotiris.nanopoulos@reddit.com>