mobile: Adds a runtime guard to drain connections on network change (#39418)
This commit adds a FALSE runtime guard
`drain_conn_pools_on_network_change` to drain network connections on
network change events when DNS refresh on network change is disabled.

The preference is to *not* refresh DNS on network change but to drain
the connections only. Right now, refreshing DNS is the only way to drain
connections too, but the goal of this commit is to decouple the two
operations and experiment with both.

Signed-off-by: Ali Beyad <abeyad@google.com>