(header_mutation): support mutating trailers (#39433)
### Commit Message:
As discussed in https://github.com/envoyproxy/envoy/issues/39382 with
`ORCA` we have the grpc upstream sending load report in the response
trailers. This can be propagated by envoy to the client which can lead
to the client load balancing against envoy based on upstream load
reports. This is not ideal in scenarios where envoy acts a gateway (e.g.
cluster ingress in k8s).

This changes extends the header mutation filter to support mutating
response trailers.

### Additional Description:
Notes on the implementation I wanted some feedback on:
- I didnt extend the `header_mutation_integration_test` because they
seemed to focus on shared functionality between the header and trailer
mutation. Happy to add it if you think it would be useful.

Risk Level: Low
Testing: unit-tests
Docs Changes: incoming
Release Notes: Incoming
Platform Specific Features: N/A
Optional Fixes #Issue
Related to #39382 but not a full fix

---------

Signed-off-by: sotiris <sotiris.nanopoulos@reddit.com>
Signed-off-by: Ting Pan <panting@google.com>