Make StringMatcher::match consistent (#39512)
Commit Message: Make StringMatcher::match consistent
Additional Description: There were instances of `StringMatcher::match`
that take `const absl::string_view` and instances that take
`absl::string_view` (not const). Making a string_view const is about as
meaningful as an argument being `const int`, in that there's little
advantage, and for a virtual function significant disadvantage (in that
an implementation may want to modify the value, and since it's passed by
value there's no reason to not allow it), so my preferred direction of
consistency is to remove the non-useful const rather than add it to the
other overrides, but I'm happy to go the other way if reviewers have
that preference.

My main reason to make it consistent is, the gcc build currently logs
over 5000 instances of warnings like
```
./envoy/common/matchers.h:34:16: warning: ‘virtual bool Envoy::Matchers::StringMatcher::match(absl::lts_20250127::string_view, const Envoy::Matchers::StringMatcher::Context&) const’ was hidden [-Woverloaded-virtual]
    34 |   virtual bool match(const absl::string_view value, const Context&) const { return match(value); }
       |                ^~~~~
 In file included from ./source/common/stats/histogram_impl.h:11,
                  from ./test/mocks/stats/mocks.h:18,
                  from ./test/mocks/api/mocks.h:20,
                  from ./test/mocks/server/server_factory_context.h:13,
                  from ./test/mocks/server/instance.h:7,
                  from ./test/integration/fake_upstream.h:41,
                  from ./test/integration/autonomous_upstream.h:3,
                  from ./test/integration/base_integration_test.h:16,
                  from ./test/integration/integration.h:5,
                  from ./test/integration/http_integration.h:16,
                  from ./test/integration/http_protocol_integration.h:3,
                  from test/extensions/filters/http/lua/lua_integration_test.cc:5:
 ./source/common/common/matchers.h:88:8: note:   by ‘virtual bool Envoy::Matchers::UniversalStringMatcher::match(absl::lts_20250127::string_view) const’
    88 |   bool match(absl::string_view) const override { return true; }
       |        ^~~~~
 In file included from ./envoy/http/codec.h:9,
                  from ./source/common/http/codec_client.h:10,
                  from ./test/integration/http_integration.h:11,
                  from ./test/integration/http_protocol_integration.h:3,
                  from test/extensions/filters/http/lua/lua_integration_test.cc:5:
```
which makes those logs barely usable. But also it's just not very good
virtual function practice!
Risk Level: No behavior change.
Testing: Does it still build?
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>
Signed-off-by: Ting Pan <panting@google.com>